Move state migrations to shared
