+++ backend/src/exercise/exercise-wrapper.ts
@@ -1,33 +1,33 @@
-import type { EntityManager } from 'typeorm';
-import { LessThan } from 'typeorm';
 import type {
     ExerciseAction,
-    StateExport,
     ExerciseIds,
-    Role,
-    UUID,
     ExerciseTimeline,
+    Role,
+    StateExport,
+    UUID,
 } from 'digital-fuesim-manv-shared';
 import {
-    ExerciseState,
-    cloneDeepMutable,
     applyAction,
-    ReducerError,
+    cloneDeepMutable,
+    ExerciseState,
     reduceExerciseState,
-    validateExerciseState,
+    ReducerError,
     validateExerciseAction,
+    validateExerciseState,
 } from 'digital-fuesim-manv-shared';
-import { IncrementIdGenerator } from '../utils/increment-id-generator';
-import { ValidationErrorWrapper } from '../utils/validation-error-wrapper';
+import type { EntityManager } from 'typeorm';
+import { LessThan } from 'typeorm';
+import { Config } from '../config';
+import type { ActionWrapperEntity } from '../database/entities/action-wrapper.entity';
 import { ExerciseWrapperEntity } from '../database/entities/exercise-wrapper.entity';
+import { migrateInDatabase } from '../database/migrate-in-database';
 import { NormalType } from '../database/normal-type';
 import type { DatabaseService } from '../database/services/database-service';
-import { Config } from '../config';
+import { pushAll, removeAll } from '../utils/array';
+import { IncrementIdGenerator } from '../utils/increment-id-generator';
 import { RestoreError } from '../utils/restore-error';
 import { UserReadableIdGenerator } from '../utils/user-readable-id-generator';
-import type { ActionWrapperEntity } from '../database/entities/action-wrapper.entity';
-import { migrateInDatabase } from '../database/state-migrations/migrations';
-import { pushAll, removeAll } from '../utils/array';
+import { ValidationErrorWrapper } from '../utils/validation-error-wrapper';
 import { ActionWrapper } from './action-wrapper';
 import type { ClientWrapper } from './client-wrapper';
 import { exerciseMap } from './exercise-map';
