+++ shared/src/index.ts
@@ -26,3 +26,4 @@ export * from './http-interfaces';
 export * from './state-helpers';
 export * from './data';
 export * from './store/action-reducers/utils';
+export * from './state-migrations';
