+++ shared/src/state-migrations/11-add-map-image-is-locked.ts
@@ -1,5 +1,6 @@
-import type { Action, UUID } from 'digital-fuesim-manv-shared';
-import type { Migration } from './migrations';
+import type { Action } from '../store';
+import type { UUID } from '../utils';
+import type { Migration } from './migration-functions';
 
 export const addMapImageIsLocked11: Migration = {
     actions: (_initialState, actions) => {
