+++ shared/src/state-migrations/9-remove-is-being-treated.ts
@@ -1,5 +1,6 @@
-import type { Action, UUID } from 'digital-fuesim-manv-shared';
-import type { Migration } from './migrations';
+import type { Action } from '../store';
+import type { UUID } from '../utils';
+import type { Migration } from './migration-functions';
 
 export const removeIsBeingTreated9: Migration = {
     actions: (_initialState, actions) => {
