+++ shared/src/state-migrations/index.ts
@@ -1,0 +1,1 @@
+export * from './migrations';
